// frontend/src/api/generated/models/sptransaction-history.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { SPTransaction } from './sptransaction'

/**
 * SP取引履歴
 * @export
 * @interface SPTransactionHistory
 */
export interface SPTransactionHistory {
  /**
   *
   * @type {Array<SPTransaction>}
   * @memberof SPTransactionHistory
   */
  transactions: Array<SPTransaction>
  /**
   *
   * @type {number}
   * @memberof SPTransactionHistory
   */
  total: number
  /**
   *
   * @type {number}
   * @memberof SPTransactionHistory
   */
  skip: number
  /**
   *
   * @type {number}
   * @memberof SPTransactionHistory
   */
  limit: number
}